src/inference/requirements.txt
python-dotenv
grpcio
grpcio-tools
redis
nest-asynico